requirements.txt
pytest==9.0.2
playwright==1.58.0
requests==2.32.5
python-dotenv==1.2.2
pydantic-settings==2.13.1
pydantic==2.12.5
allure-pytest==2.15.3
jsonschema==4.26.0
ruff==0.15.7
pre-commit==4.5.1
